webfrontend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Babel==4.0.0
ldap3==2.9.1
PyYAML==6.0.1
cryptography==41.0.4
Werkzeug==3.0.0
SQLAlchemy==2.0.21
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.2
Babel==2.13.0
pytz==2023.3.post1